Add files via upload
I wrote this script as an excercise. The idea was to write an R script which, as an output, would produce a visual representation of the positions on the guitar neck corresponding to the notes of a given scale in a given key. The system is based on a look-up table approach and can easily be extended by the user to include other scales not already present in the "scale dictionary" already present in the script.